Make telemetry optional.
